webgui: extract scheme handler in rootqt5 into separate class
These classes can be reused later in other application.
One could try to generate qt library for them, but for the moment it is
enough to have source code here
