+++ gui/webdisplay/qt5/rootqt5.pro
@@ -3,10 +3,12 @@ TEMPLATE = app
 QT += webengine webenginewidgets
 
 HEADERS += rootwebpage.h \
-           rootwebview.h
+           rootwebview.h \
+           rooturlschemehandler.h
 
 SOURCES += rootwebpage.cpp \
            rootwebview.cpp \
+           rooturlschemehandler.cpp \
            rootqt5.cpp
 
 DESTDIR = $$system(root-config --bindir)
